backend/ai-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.1
tree-sitter-typescript==0.20.2
tree-sitter-go==0.20.0
tree-sitter-cpp==0.20.0

# Database and storage
psycopg2-binary==2.9.9
redis==5.0.1
pgvector==0.2.3

# Processing and utilities
numpy==1.25.2
pandas==2.1.3
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring
sentry-sdk==1.38.0

# Graph visualization
graphviz==0.20.1
matplotlib==3.8.2
